2008-2024
hard to believe that i've dreamt *and* recorded about twice this many dreams in my lifetime (2001-2007 or 2003-2007; these represent 2008-2024, but i didn't record many dreams in 2016, lost half the ones i did record; didn't record many in 2017 and lost all of the ones i did; didn't record many in 2018 and lost half of those; in 2019 i didn't record any, none in 2020, and none in 2021 or 2022; i only started again at the very end of a period of homelessness at the end of 2023 when i felt grateful to be sleeping in a bed again; i'm sad that i let the process go for so many years. it's an important one. i really wish i could re-do my life strictly around the idea that losing out on dreams means losing out on soul.)

it's also annoying that i recorded those other dreams on paper! so, so annoying.

when you write on paper, it's impossible to really own what you write. for everything i hate about the internet, being able to preserve one's writing is a huge benefit.

i had to spend a while hunting for the right file to one-time pad this.

i would have thought that the classic Chinese novels, Hong Lou Meng, or Shui Hu Zhuan, had enough bytes to pad.

i was wrong!

it's fascinating to look at literature through that kind of utilitarian lens, for sure.